requirements.txt
transformers
Pillow
gradio
numpy
torch
torchvision
timm
seaborn
opencv-python